+++ redis_server/src/data/storage_service/operator_service/request_message.rs
@@ -56,5 +56,5 @@ pub enum StorageAction {
     DecrBy(String, i32),
     IncrBy(String, i32),
     Touch(String),
-    Expire(String, i32),
+    Expire(String, u128),
 }
